requirements.txt
gradio==5.12.0
torch>=2.0.0
transformers>=4.31.0
python-dotenv>=1.0.0
edge-tts>=6.1.5
librosa>=0.10.1
bs4>=0.0.1
requests>=2.31.0
accelerate>=0.26.0
duckduckgo-search